replica: Expose table_load_stats in table
This is the table replica state that coordinator will aggregate
from all nodes and feed into the load balancer.

A tablet filter is added to not double account migrating tablets,
so only one of pending or leaving tablet replica will be accounted
based on current migration stage. More details can be known in
the patch that will implement the filter.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>