test/topology: add --request-timeout option
the default 200 second request_timeout might be too long in the manual
tests. for instance, when the test completes, the python driver could
still be trying to establish connections to the server, and when we
shutdown the driver when teardown the test, the driver blocks on the
connection in vain for up to 200 seconds, as the cluster is already
down, until timeout expires. this test cycle is quite long.

so, in this change, an other option is added to configure the
timeout, so one can override the default 200 seconds by changing
the per test's pytest.ini like
```ini
[pytest]
addopts = --request-timeout=10
```

the existing behavior is preserved, so the default value of
`--request-timeout` is still 200.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>