pylib/scylla_cluster: return an iterator of log messages
before this change, `ScyllaCluster.read_server_log()` and
`ScyllaCluster.server_log_filename()` return a `str` and
a path to the first server's logging message. this is a known
limit of these two functions.

because a cluster could contain multiple servers, and the first
server's logging message might not enough for the postmortem
debugging. to prepare for the next change which prints all logging
messages of all servers, let's return an iterator instead.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>