test.py: print out paths to all servers log for TopologyTest
TopologyTest is likely to involve multiple servers. so it would be
ideal if we can print out all servers' logging messages when the
test fails.

before this change,

- `TopologyTest` does not collect the server's logging message
  if a test fails. it does so only when an exception is thrown.
  but if a test fails without throwing an exception, the server
  logging messages are not collected.
- `TopologyTest` just inherits the behavior of
  `PythonTest.print_summary()`, where only a single server's log
  messagess are printed when an exception is thrown.

after this change,

- the paths to the servers' logging files are collected when
  the test fails or throws.
- `TopologyTest` overrides `print_summary()` and print out
  paths to all servers' logging messages if the test fails.

`self.server_log_filename` is used when printing the
summary of JUnit tests in `write_junit_report()`, which is used
when printing the test summary in the JUnit test report XML.
and this XML is in turn consumed by Jenkins for rendering on
a per-test basis. since we already have the paths printed already
stdout, there is no need to include the path of the first
server in the XML file, so the corresponding
`TopologyTest.write_junit_failure_report()` is not defined.

Refs #16656
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>