test/cql-pytest: add reproducer for task-tracking memory leak
This patch adds a reproducer test for the memory leak described in
issue #16493: If a table is repeatedly created and dropped, memory
is leaked by task tracking. The problem is severely exacerbated when
tablets are used which is why the test here uses tablets.

Before the fix for #16493 (a Seastar patch, scylladb/seastar#2023),
this test of 100 iterations always failed (with test/cql-pytest/run's
default memory allowance). After the fix (in the previous commit),
the test doesn't fail in 100 iterations - and even if increased to
10,000 iterations it doesn't fail.

The new test uses the initial_tablets feature, so requires Scylla to be
run with the "tablets" experimental option turned on. This is not
currently the default of test.py or test/cql-pytest/run, so I turned
it on manually to check this test. I also checked that the test is
correctly skipped if tablets are not turned on.

Refs #16493

Signed-off-by: Nadav Har'El <nyh@scylladb.com>