streaming: handle no_such_column_family gracefully
Schema propagation is asynchronous, so a sender node may not know yet
that a table was dropped and stream it, while receiver has already
dropped	the table.

Currently in this situation receiver may reply with no_such_column_family
exception. It is not possible to properly interpret that on the other
node (no_such_column_family should be indicated by returning -2 instead
of exception) and so it closes the session, causing streaming to fail.

To prevent that, enclose find_column_family (which may throw) call
in register_stream_mutation_fragments in lambda so that it can be handled
in the background, the same way as latter thrown find_column_family are.