alternator: avoid assertion failure when configured with a busy port
If Alternator is configured to use a port which is already busy (e.g.,
port 9042 which is used by CQL), we expect a clean error message like:

  ERROR 2024-01-10 18:32:44,948 [shard 0:main] init - Startup failed:
  std::system_error (error system:98, posix_listen failed for address
  127.0.0.1:9042: Address already in use)

This indeed happens when running Scylla on a *single core*. But when we
have multiple cores, we encounter a Seastar bug scylladb/seastar#2036:
Core 0 fails its listen cleanly as expected (with an exception as shown
above), but all other cores fail on a nasty assertion failure:

  seastar/src/net/posix-stack.cc:553: virtual future<accept_result>
  seastar::net::posix_ap_server_socket_impl::accept():
  Assertion `i.second' failed.

This causes Scylla to crash in an ugly way, and the user doesn't
understand what really went wrong.

The solution in this patch works around the Seastar bug by first
running the server initialization function on core 0, and only
then running it on all other cores. The port already in use is
detected on shard 0 and we never try to do it on the other shards
and don't reach the assertion failure.

This solution is also better that if in the future the http server
initialization will have other configuration checks: It's better
to get these errors once from shard 0 - and not in parallel from
all shards.

Fixes #16720.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>