main: fix "starting {}" messages
The supervisor::notify() function expects a single string - not a
format and parameters. Calls we have in main.cc like

    supervisor::notify("starting {}", what);

end up printing the silly message "starting {}". The second parameter
"what" is converted to a bool, also having an unintended consequence
for telling notify we're "ready".

This patch fixes it to call fmt::format, as intended.

Fixes #16728

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

Closes scylladb/scylladb#16729